+++ lib/src/node_blue_print.dart
@@ -102,7 +102,11 @@ class NodeBluePrint<T> {
   }
 
   /// Maps the key of the blue print to another key
-  NodeBluePrint<T> mapTo(String key) => copyWith(key: key);
+  NodeBluePrint<T> forward({required String toKey}) => NodeBluePrint.map(
+        supplier: key,
+        toKey: toKey,
+        initialProduct: initialProduct,
+      );
 
   /// Provites an operator =
   @override